+++ CHANGELOG.md
@@ -1,6 +1,20 @@
-## Unreleased
+## 0.22.0 (2023-11-09)
+
+### Breaking changes
 
 - Use `{providedIn: 'root'}` for Angular endpoints. This facilitates tree-shaking of unused code.
+- Don't create `ApinaModule` for Angular, but a `provideApina`-function instead that you can use to create.
+  a provider. Also, `HttpClient` is not automatically imported anymore, you should use Angular's `provideHttpClient`
+  to register a client. To migrate, remove `ApinaModule from imports` and use this instead:
+    ```typescript
+    {
+        providers: [
+           provideHttpClient(),
+           provideApina()
+        ]
+    }
+    ```
+  To customize Apina configuration, don't add additional providers manually, but pass parameters to `provideApina`.
 
 ## 0.21.0 (2023-08-02)